Extract SetupForm specific instructions to new component.
