+++ assets/js/modules/tagmanager/components/setup/index.js
@@ -18,3 +18,4 @@
 
 export { default as SetupMain } from './SetupMain';
 export { default as SetupForm } from './SetupForm';
+export { default as SetupFormInstructions } from './SetupFormInstructions';
